+++ x/ugov/keeper/query_server.go
@@ -24,3 +24,11 @@ func (q Querier) MinGasPrice(ctx context.Context, _ *ugov.QueryMinGasPrice) (*ug
 	return &ugov.QueryMinGasPriceResponse{MinGasPrice: q.Keeper(&sdkCtx).MinGasPrice()},
 		nil
 }
+
+func (q Querier) EmergencyGroup(ctx context.Context, _ *ugov.QueryEmergencyGroup,
+) (*ugov.QueryEmergencyGroupResponse, error) {
+	sdkCtx := sdk.UnwrapSDKContext(ctx)
+	return &ugov.QueryEmergencyGroupResponse{
+			EmergencyGroup: q.Keeper(&sdkCtx).EmergencyGroup().String()},
+		nil
+}